requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bleach==5.0.1
Django==4.0.6
django-cuser==2017.3.16
django-summernote==0.8.20.0
django-taggit==3.0.0
django-tinymce==3.4.0
modeltranslation==0.25
Pillow==8.1.1
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
